renamed files and folders to use snake-case
