+++ src/pages/my-list/my-list.tsx
@@ -1,7 +1,7 @@
-import FilmsList from '../../components/films/filmsList';
+import FilmsList from '../../components/films/films-list';
 import Footer from '../../components/footer/footer';
 import Logo from '../../components/header/logo';
-import UserBlock from '../../components/header/userBlock';
+import UserBlock from '../../components/header/user-block';
 import { myListFilms } from '../../data';
 
 function MyList(): JSX.Element {
